refactor(webapp): split monolith into storage/connectors/history/runtime/api modules
Break angelus/webapp.py (2373 lines, 91 functions) into focused modules:
- storage.py: state root, session registry, event log, in-memory locks
- connectors.py: RSA-OAEP encrypted connector store
- history.py: transcript/archive/usage rebuilds + legacy migration
- runtime.py: Agent/Swarm construction and run-scoped helpers
- markdown.py: bounded markdown-it rendering shared by the above
- api/: FastAPI routers split into connectors/sessions/runs

webapp.py is now a thin assembly shell that re-exports every symbol so the
CLI (angelus.webapp.app / WORKSPACE_ROOT / _read_workspaces), uvicorn
(angelus.webapp:app) and the regression suite keep working unchanged.

Tests that mutate module-level state now target the owning module
(storage.WORKSPACE_ROOT etc) instead of the webapp facade.
Full suite: 208 passed.
