+++ angelus/markdown.py
@@ -1,0 +1,39 @@
+"""Bounded Markdown rendering shared by the web API and history rebuilds."""
+
+from __future__ import annotations
+
+from collections import OrderedDict
+
+from markdown_it import MarkdownIt
+
+# Raw HTML remains disabled so model output cannot inject arbitrary browser DOM.
+_MARKDOWN = MarkdownIt("commonmark", {"html": False, "linkify": False}).enable("table")
+_MARKDOWN_CACHE_MAX = 512
+_markdown_cache: "OrderedDict[str, str]" = OrderedDict()
+
+def render_markdown(text: str) -> str:
+    """Convert trusted-to-render model Markdown into safe display HTML.
+
+    Rendering long transcripts on every history reload is expensive.  A small
+    bounded cache (LRU by insertion order) avoids re-rendering unchanged
+    assistant turns across refreshes, session switches, and event-log rebuilds.
+
+    Args:
+        text: Raw assistant or reasoning text received from an LLM backend.
+
+    Returns:
+        HTML generated by ``markdown-it-py`` with raw HTML and linkification
+        disabled. The caller may insert this value into the console message UI.
+    """
+    cached = _markdown_cache.get(text)
+    if cached is not None:
+        _markdown_cache.move_to_end(text)
+        return cached
+    html = _MARKDOWN.render(text)
+    _markdown_cache[text] = html
+    if len(_markdown_cache) > _MARKDOWN_CACHE_MAX:
+        _markdown_cache.popitem(last=False)
+    return html
+
+__all__ = ["render_markdown", "_MARKDOWN", "_MARKDOWN_CACHE_MAX", "_markdown_cache"]
+
