+++ tests/test_run_profile_persistence.py
@@ -10,7 +10,7 @@ from types import SimpleNamespace
 from unittest.mock import patch
 from pathlib import Path
 
-from angelus import webapp
+from angelus import runtime, storage, webapp
 from angelus.classes import RunConfig
 from angelus.classes import RunRequest
 from llmfetcher.llm_types import LLMOutput
@@ -67,8 +67,8 @@ class RunProfilePersistenceTests(unittest.TestCase):
     def test_concurrent_event_appends_remain_complete_json_records(self) -> None:
         """Swarm worker hooks must not interleave their NDJSON records."""
         with tempfile.TemporaryDirectory() as directory:
-            original_root = webapp.WORKSPACE_ROOT
-            webapp.WORKSPACE_ROOT = Path(directory)
+            original_root = storage.WORKSPACE_ROOT
+            storage.WORKSPACE_ROOT = Path(directory)
             try:
                 workers = [
                     threading.Thread(
@@ -86,24 +86,24 @@ class RunProfilePersistenceTests(unittest.TestCase):
                 self.assertEqual(len(events), 32)
                 self.assertEqual({event["index"] for event in events}, set(range(32)))
             finally:
-                webapp.WORKSPACE_ROOT = original_root
+                storage.WORKSPACE_ROOT = original_root
 
     def test_start_run_persists_profile_in_state_and_event_log(self) -> None:
         """Run provenance survives both the active and terminal state rewrite."""
         with tempfile.TemporaryDirectory() as directory:
-            original_root = webapp.WORKSPACE_ROOT
+            original_root = storage.WORKSPACE_ROOT
             key = ("demo", "demo")
-            webapp.WORKSPACE_ROOT = Path(directory)
-            with webapp._sessions_lock:
-                prior = webapp._sessions.pop(key, None)
+            storage.WORKSPACE_ROOT = Path(directory)
+            with storage._sessions_lock:
+                prior = storage._sessions.pop(key, None)
             request = RunRequest(
                 workspace_id="demo", session_id="demo", message="hello",
                 config=RunConfig(model="test", api_key="hidden", system_prompt="private"),
             )
             try:
                 with (
-                    patch.object(webapp, "_workspace_exists", return_value=True),
-                    patch.object(webapp, "_build_agent", return_value=_CompletedAgent()),
+                    patch.object(storage, "_workspace_exists", return_value=True),
+                    patch.object(runtime, "_build_agent", return_value=_CompletedAgent()),
                     patch.object(webapp.threading, "Thread", _ImmediateThread),
                 ):
                     webapp.start_run(request)
@@ -118,8 +118,8 @@ class RunProfilePersistenceTests(unittest.TestCase):
                     state["runtime_profile"]["fingerprint"],
                 )
             finally:
-                webapp.WORKSPACE_ROOT = original_root
-                with webapp._sessions_lock:
-                    webapp._sessions.pop(key, None)
+                storage.WORKSPACE_ROOT = original_root
+                with storage._sessions_lock:
+                    storage._sessions.pop(key, None)
                     if prior is not None:
-                        webapp._sessions[key] = prior
+                        storage._sessions[key] = prior
